requirements.txt
pytorch-argus==1.0.0
numpy==1.26.2
scipy==1.11.4
opencv-python-headless==4.8.1.78
scikit-learn==1.3.2
matplotlib==3.8.2
notebook==7.0.6
hydra-core==1.3.2
optuna==2.10.1
hydra-optuna-sweeper==1.2.0
# Dev tools
tqdm==4.66.1
mypy==1.7.0
pytest==7.4.3
